flipay v0.1.0 Flipay.BestRateFinder

Find the best rate according to input/output assets, input amount and exchange's order book.

Functions

buy_best_rate(order_books, remain_amount, total_size)

Calculates the best buying rate according to quotes and input amount.

Examples

iex> order_books = [%Flipay.Exchanges.Quote{number_of_order: 1, price: 5000, size: 1}, %Flipay.Exchanges.Quote{number_of_order: 2, price: 5100, size: 1}]
iex> {:ok, size} = Flipay.BestRateFinder.buy_best_rate(order_books, 10100, 0)
iex> size
#Decimal<2>
iex> {:ok, size} = Flipay.BestRateFinder.buy_best_rate(order_books, 15200, 0)
iex> size
#Decimal<3>
iex> {:ok, size} = Flipay.BestRateFinder.buy_best_rate(order_books, 15000, 0)
iex> size
#Decimal<2.960784313725490196078431373>
iex> Flipay.BestRateFinder.buy_best_rate(order_books, 16000, 0)
{:error, :not_enough_quotes}

find(map)

Finds the best rate for input request. Order book comes from specific exchange and the quotes are sorted by best to worst order. Input/output assets and input amount are specified by user.

Examples:

iex> order_book = %Flipay.Exchanges.OrderBook{
...>  exchange: Flipay.Exchanges.Coinbase,
...>  exchange_side: "asks",
...>  input_asset: "USD",
...>  output_asset: "BTC",
...>  quotes: [
...>    %Flipay.Exchanges.Quote{
...>      number_of_order: 1,
...>      price: 5000,
...>      size: 2
...>      },
...>    %Flipay.Exchanges.Quote{
...>      number_of_order: 1,
...>      price: 6000,
...>      size: 1
...>    }
...>  ]
...>}
iex> {:ok, rate} = Flipay.BestRateFinder.find(%{order_book: order_book, input_amount: 12000})
iex> rate
#Decimal<2.333333333333333333333333333>

sell_best_rate(quotes, remain_size, total_amount)

Calculates the best selling rate according to quotes and input size.

Examples:

iex> quotes = [%Flipay.Exchanges.Quote{number_of_order: 1, price: 5000, size: 1}, %Flipay.Exchanges.Quote{number_of_order: 2, price: 4900, size: 1}]
iex> {:ok, amount} = Flipay.BestRateFinder.sell_best_rate(quotes, 1, 0)
iex> amount
#Decimal<5000>
iex> {:ok, input_size} = Decimal.parse("1.5")
iex> {:ok, amount} = Flipay.BestRateFinder.sell_best_rate(quotes, input_size, 0)
iex> amount
#Decimal<7450.0>
iex> {:ok, input_size} = Decimal.parse("3.1")
iex> Flipay.BestRateFinder.sell_best_rate(quotes, input_size, 0)
{:error, :not_enough_quotes}