flipay v0.1.0 Flipay.Exchanges

Get order book from specific exchange, according to input/output assets.

Functions

convert_to_decimal(order_book)

Gets quotes from order books by specific exchange side.

Examples

iex> input = %Flipay.Exchanges.OrderBook{quotes: %{ "5000" => "2" }}
iex> order_book = Flipay.Exchanges.convert_to_decimal(input)
iex> {:ok, quote} = Enum.fetch(order_book.quotes,0)
iex> quote.price
#Decimal<5000>
iex> quote.size
#Decimal<2>

get_asset_type(asset)

Gets the type of specific asset.

Examples

iex> Flipay.Exchanges.get_asset_type("TWD")
{:error, :unsupported_asset}

iex> Flipay.Exchanges.get_asset_type("USD")
{:ok, :fiat_money}

iex(3)> Flipay.Exchanges.get_asset_type("BTC")
{:ok, :digital_currency}

get_exchange(exchange_name)

Get exchange module by exchange name.

Examples:

iex> Flipay.Exchanges.get_exchange("coinbase_pro")
{:ok,
%Flipay.Exchanges.OrderBook{
  exchange: Flipay.Exchanges.Coinbase,
  exchange_side: nil,
  input_asset: nil,
  output_asset: nil,
  quotes: nil
}}
iex> Flipay.Exchanges.get_exchange("hitbtc")
{:error, :not_found}

get_order_book(order_book)

Get order books from exchange.

Examples:

iex> order_book = %Flipay.Exchanges.OrderBook{exchange: Flipay.Exchanges.Coinbase, asset: "BTC-USD", exchange_side: "asks"}
iex> {:ok, order_book} = Flipay.Exchanges.get_order_book(order_book)
iex> map_size(order_book.quotes) > 0
true

get_quotes(map)

Get quotes by specific exchange name and input/output assets.

Examples:

iex> input = %{exchange_name: "coinbase", input_asset: "USD", output_asset: "BTC"}
iex> {:ok, order_book} = Flipay.Exchanges.get_quotes(input)
iex> Enum.count(order_book.quotes) > 0
true
iex> input = %{input | exchange_name: "hitbtc"}
iex> Flipay.Exchanges.get_quotes(input)
{:error, :not_found}
iex> input = %{input | exchange_name: "coinbase", output_asset: "TWD"}
iex> Flipay.Exchanges.get_quotes(input)
{:error, :unsupported_asset}

set_assets(order_book, input_asset, output_asset)

Sets input/output assets and determine the exchange side.

Examples:

iex> order_book = %Flipay.Exchanges.OrderBook{}
iex> Flipay.Exchanges.set_assets(order_book, "USD", "BTC")
{:ok,
%Flipay.Exchanges.OrderBook{
  exchange: nil,
  asset: "BTC-USD",
  exchange_side: "asks",
  quotes: nil
}}
iex> Flipay.Exchanges.set_assets(order_book, "ETH", "USD")
{:ok,
%Flipay.Exchanges.OrderBook{
  exchange: nil,
  asset: "ETH-USD",
  exchange_side: "bids",
  quotes: nil
}}
iex> Flipay.Exchanges.set_assets(order_book, "ETH", "TWD")
{:error, :unsupported_asset}
iex> Flipay.Exchanges.set_assets(order_book, "TWD", "USD")
{:error, :unsupported_asset}

sort_quotes(order_book)

Sorts the quotes by exchange side to make sure the quotes are sorted from best to worst.

Examples:

iex> order_book = %Flipay.Exchanges.OrderBook{quotes: [%{price: 5000}, %{price: 4900}, %{price: 5100}], exchange_side: "asks"}
iex> Flipay.Exchanges.sort_quotes(order_book)
%Flipay.Exchanges.OrderBook{
  exchange: nil,
  exchange_side: "asks",
  input_asset: nil,
  output_asset: nil,
  quotes: [%{price: 4900}, %{price: 5000}, %{price: 5100}]
}
iex> order_book = %Flipay.Exchanges.OrderBook{order_book | exchange_side: "bids"}
iex> Flipay.Exchanges.sort_quotes(order_book)
%Flipay.Exchanges.OrderBook{
  exchange: nil,
  exchange_side: "bids",
  input_asset: nil,
  output_asset: nil,
  quotes: [%{price: 5100}, %{price: 5000}, %{price: 4900}]
}